requirements.txt
fastapi
uvicorn[standard]
requests
httpx
beautifulsoup4
demjson3
selenium
psycopg2-binary
sqlalchemy
python-dotenv